Moved tests to seperate folder at root or yarn dev fails
